De-duplicate spec tests code across forks
