+++ packages/spec-test-runner/test/spec/altair/epoch_processing.test.ts
@@ -1,19 +1,11 @@
-import {join} from "path";
-import fs from "fs";
-
-import {CachedBeaconState, allForks, altair} from "@chainsafe/lodestar-beacon-state-transition";
-import {describeDirectorySpecTest} from "@chainsafe/lodestar-spec-test-util";
+import {allForks, altair} from "@chainsafe/lodestar-beacon-state-transition";
 import {processParticipationRecordUpdates} from "@chainsafe/lodestar-beacon-state-transition/src/phase0/epoch/processParticipationRecordUpdates";
-import {altair as altairTypes, ssz} from "@chainsafe/lodestar-types";
-import {TreeBacked} from "@chainsafe/ssz";
-import {ACTIVE_PRESET, ForkName} from "@chainsafe/lodestar-params";
-import {SPEC_TEST_LOCATION} from "../../specTestVersioning";
-import {expectEqualBeaconState, inputTypeSszTreeBacked} from "../util";
-import {IAltairStateTestCase, config} from "./util";
+import {ForkName} from "@chainsafe/lodestar-params";
+import {EpochProcessFn, epochProcessing} from "../allForks/epochProcessing";
 
-/** Describe with which function to run each directory of tests */
-const epochProcessFns: Record<string, EpochProcessFn> = {
-  /* eslint-disable @typescript-eslint/naming-convention */
+/* eslint-disable @typescript-eslint/naming-convention */
+
+epochProcessing(ForkName.altair, {
   effective_balance_updates: allForks.processEffectiveBalanceUpdates,
   eth1_data_reset: allForks.processEth1DataReset,
   historical_roots_update: allForks.processHistoricalRootsUpdate,
@@ -27,37 +19,4 @@ const epochProcessFns: Record<string, EpochProcessFn> = {
   slashings: altair.processSlashings as EpochProcessFn,
   slashings_reset: allForks.processSlashingsReset,
   sync_committee_updates: altair.processSyncCommitteeUpdates as EpochProcessFn,
-  /* eslint-enable @typescript-eslint/naming-convention */
-};
-type EpochProcessFn = (state: CachedBeaconState<allForks.BeaconState>, epochProcess: allForks.IEpochProcess) => void;
-
-const rootDir = join(SPEC_TEST_LOCATION, `tests/${ACTIVE_PRESET}/altair/epoch_processing`);
-for (const testDir of fs.readdirSync(rootDir)) {
-  const epochProcessFn = epochProcessFns[testDir];
-  if (!epochProcessFn) {
-    throw Error(`No epochProcessFn for ${testDir}`);
-  }
-
-  describeDirectorySpecTest<IAltairStateTestCase, altairTypes.BeaconState>(
-    `${ACTIVE_PRESET}/altair/epoch_processing/${testDir}`,
-    join(rootDir, `${testDir}/pyspec_tests`),
-    (testcase) => {
-      const stateTB = (testcase.pre as TreeBacked<allForks.BeaconState>).clone();
-      const state = allForks.createCachedBeaconState(config, stateTB);
-      const epochProcess = allForks.beforeProcessEpoch(state);
-      epochProcessFn(state, epochProcess);
-      return state;
-    },
-    {
-      inputTypes: inputTypeSszTreeBacked,
-      sszTypes: {
-        pre: ssz.altair.BeaconState,
-        post: ssz.altair.BeaconState,
-      },
-      getExpected: (testCase) => testCase.post,
-      expectFunc: (testCase, expected, actual) => {
-        expectEqualBeaconState(ForkName.altair, expected, actual);
-      },
-    }
-  );
-}
+});
