+++ packages/spec-test-runner/test/spec/phase0/operations.test.ts
@@ -1,19 +1,12 @@
-import {join} from "path";
-import fs from "fs";
 import {CachedBeaconState, allForks, phase0} from "@chainsafe/lodestar-beacon-state-transition";
-import {describeDirectorySpecTest} from "@chainsafe/lodestar-spec-test-util";
-import {ssz} from "@chainsafe/lodestar-types";
-import {TreeBacked} from "@chainsafe/ssz";
-import {ACTIVE_PRESET, ForkName} from "@chainsafe/lodestar-params";
-import {SPEC_TEST_LOCATION} from "../../specTestVersioning";
-import {expectEqualBeaconState, inputTypeSszTreeBacked} from "../util";
-import {IPhase0StateTestCase, config} from "./util";
+import {ForkName} from "@chainsafe/lodestar-params";
 import {IBaseSpecTest} from "../type";
+import {operations} from "../allForks/operations";
 
 /* eslint-disable @typescript-eslint/naming-convention */
 
 /** Describe with which function to run each directory of tests */
-const operationFns: Record<string, EpochProcessFn> = {
+operations<phase0.BeaconState>(ForkName.phase0, {
   attestation: (state, testCase: IBaseSpecTest & {attestation: phase0.Attestation}) => {
     phase0.processAttestation(state, testCase.attestation);
   },
@@ -38,42 +31,4 @@ const operationFns: Record<string, EpochProcessFn> = {
   voluntary_exit: (state, testCase: IBaseSpecTest & {voluntary_exit: phase0.SignedVoluntaryExit}) => {
     phase0.processVoluntaryExit(state, testCase.voluntary_exit);
   },
-};
-type EpochProcessFn = (state: CachedBeaconState<phase0.BeaconState>, testCase: any) => void;
-
-const rootDir = join(SPEC_TEST_LOCATION, `tests/${ACTIVE_PRESET}/phase0/operations`);
-for (const testDir of fs.readdirSync(rootDir)) {
-  const operationFn = operationFns[testDir];
-  if (!operationFn) {
-    throw Error(`No operationFn for ${testDir}`);
-  }
-
-  describeDirectorySpecTest<IPhase0StateTestCase, phase0.BeaconState>(
-    `${ACTIVE_PRESET}/phase0/operations/${testDir}`,
-    join(rootDir, `${testDir}/pyspec_tests`),
-    (testcase) => {
-      const stateTB = (testcase.pre as TreeBacked<phase0.BeaconState>).clone();
-      const state = allForks.createCachedBeaconState(config, stateTB);
-      operationFn(state, testcase);
-      return state;
-    },
-    {
-      inputTypes: inputTypeSszTreeBacked,
-      sszTypes: {
-        pre: ssz.phase0.BeaconState,
-        post: ssz.phase0.BeaconState,
-        attestation: ssz.phase0.Attestation,
-        attester_slashing: ssz.phase0.AttesterSlashing,
-        block: ssz.phase0.BeaconBlock,
-        deposit: ssz.phase0.Deposit,
-        proposer_slashing: ssz.phase0.ProposerSlashing,
-        voluntary_exit: ssz.phase0.SignedVoluntaryExit,
-      },
-      shouldError: (testCase) => !testCase.post,
-      getExpected: (testCase) => testCase.post,
-      expectFunc: (testCase, expected, actual) => {
-        expectEqualBeaconState(ForkName.phase0, expected, actual);
-      },
-    }
-  );
-}
+});
